+++ Library/OcCpuLib/AppleCpuSupport.c
@@ -658,3 +658,31 @@ OcCpuModelToAppleFamily (
       return CPUFAMILY_UNKNOWN;
   }
 }
+
+UINT16
+OcCpuFrequencyToDisplayFrequency (
+  IN UINT64  Frequency
+  )
+{
+  UINT16                          MhzSpeed;
+  UINT16                          MhzRemainder;
+
+  //
+  // Round to nearest in MHz
+  //
+  MhzSpeed = (UINT16) DivU64x32 (Frequency + 500000, 1000000);
+  MhzRemainder = MhzSpeed % 100;
+  //
+  // Round to two digits when the second digit is above zero or to one otherwise.
+  // REF: https://github.com/acidanthera/bugtracker/issues/1521
+  //
+  if (MhzRemainder >= 60 && MhzRemainder < 89) {
+    MhzSpeed = (MhzSpeed) / 10 * 10;
+  } else if (MhzRemainder >= 12 && MhzRemainder < 89) {
+    MhzSpeed = (MhzSpeed + 5) / 10 * 10;
+  } else {
+    MhzSpeed = (MhzSpeed + 50) / 100 * 100;
+  }
+
+  return MhzSpeed;
+}