+++ Platform/OpenCanopy/Views/BootPicker.c
@@ -1298,24 +1298,9 @@ GLOBAL_REMOVE_IF_UNREFERENCED GUI_OBJ_CLICKABLE mBootPickerRightScroll = {
   NULL
 };
 
-GLOBAL_REMOVE_IF_UNREFERENCED GUI_OBJ_CLICKABLE mBootPickerShutDown = {
-  {
-    { &mBootPickerActionButtonsContainer.Obj.Children, &mBootPickerRestart.Hdr.Link },
-    &mBootPickerActionButtonsContainer.Obj,
-    {
-      0, 0, 0, 0,
-      InternalBootPickerSimpleButtonDraw,
-      InternalBootPickerShutDownPtrEvent,
-      NULL,
-      INITIALIZE_LIST_HEAD_VARIABLE (mBootPickerShutDown.Hdr.Obj.Children)
-    }
-  },
-  NULL
-};
-
 GLOBAL_REMOVE_IF_UNREFERENCED GUI_OBJ_CLICKABLE mBootPickerRestart = {
   {
-    { &mBootPickerShutDown.Hdr.Link, &mBootPickerActionButtonsContainer.Obj.Children },
+    { &mBootPickerActionButtonsContainer.Obj.Children, &mBootPickerShutDown.Hdr.Link },
     &mBootPickerActionButtonsContainer.Obj,
     {
       0, 0, 0, 0,
@@ -1323,6 +1308,21 @@ GLOBAL_REMOVE_IF_UNREFERENCED GUI_OBJ_CLICKABLE mBootPickerRestart = {
       InternalBootPickerRestartPtrEvent,
       NULL,
       INITIALIZE_LIST_HEAD_VARIABLE (mBootPickerRestart.Hdr.Obj.Children)
+    }
+  },
+  NULL
+};
+
+GLOBAL_REMOVE_IF_UNREFERENCED GUI_OBJ_CLICKABLE mBootPickerShutDown = {
+  {
+    { &mBootPickerRestart.Hdr.Link, &mBootPickerActionButtonsContainer.Obj.Children },
+    &mBootPickerActionButtonsContainer.Obj,
+    {
+      0, 0, 0, 0,
+      InternalBootPickerSimpleButtonDraw,
+      InternalBootPickerShutDownPtrEvent,
+      NULL,
+      INITIALIZE_LIST_HEAD_VARIABLE (mBootPickerShutDown.Hdr.Obj.Children)
     }
   },
   NULL
@@ -1336,7 +1336,7 @@ GLOBAL_REMOVE_IF_UNREFERENCED GUI_OBJ_CHILD mBootPickerActionButtonsContainer = 
     GuiObjDrawDelegate,
     GuiObjDelegatePtrEvent,
     NULL,
-    { &mBootPickerRestart.Hdr.Link, &mBootPickerShutDown.Hdr.Link }
+    { &mBootPickerShutDown.Hdr.Link, &mBootPickerRestart.Hdr.Link }
   }
 };
 
@@ -1833,25 +1833,29 @@ BootPickerViewInitialize (
   // padding.
   //
   mBootPicker.Hdr.Obj.Width   = 0U - (UINT32) (BOOT_ENTRY_SPACE * GuiContext->Scale);
-  mBootPicker.Hdr.Obj.OffsetX = mBootPickerContainer.Obj.Width / 2;
+  //
+  // Adding an entry will also shift OffsetX considering the added boot entry
+  // space. This is not needed for the first, so initialise accordingly.
+  //
+  mBootPicker.Hdr.Obj.OffsetX = mBootPickerContainer.Obj.Width / 2 + (UINT32) (BOOT_ENTRY_SPACE * GuiContext->Scale) / 2;
   mBootPicker.Hdr.Obj.OffsetY = 0;
 
   mBootPicker.SelectedEntry = NULL;
 
+  mBootPickerRestart.CurrentImage = &GuiContext->Icons[ICON_RESTART][ICON_TYPE_BASE];
+  mBootPickerRestart.Hdr.Obj.Width = mBootPickerRestart.CurrentImage->Width;
+  mBootPickerRestart.Hdr.Obj.Height = mBootPickerRestart.CurrentImage->Height;
+  mBootPickerRestart.Hdr.Obj.OffsetX = 0;
+  mBootPickerRestart.Hdr.Obj.OffsetY = 0;
+
   mBootPickerShutDown.CurrentImage = &GuiContext->Icons[ICON_SHUT_DOWN][ICON_TYPE_BASE];
   mBootPickerShutDown.Hdr.Obj.Width = mBootPickerShutDown.CurrentImage->Width;
   mBootPickerShutDown.Hdr.Obj.Height = mBootPickerShutDown.CurrentImage->Height;
-  mBootPickerShutDown.Hdr.Obj.OffsetX = 0;
+  mBootPickerShutDown.Hdr.Obj.OffsetX = mBootPickerRestart.Hdr.Obj.Width + BOOT_ACTION_BUTTON_SPACE * GuiContext->Scale;
   mBootPickerShutDown.Hdr.Obj.OffsetY = 0;
 
-  mBootPickerRestart.CurrentImage = &GuiContext->Icons[ICON_RESTART][ICON_TYPE_BASE];
-  mBootPickerRestart.Hdr.Obj.Width = mBootPickerShutDown.CurrentImage->Width;
-  mBootPickerRestart.Hdr.Obj.Height = mBootPickerShutDown.CurrentImage->Height;
-  mBootPickerRestart.Hdr.Obj.OffsetX = mBootPickerShutDown.Hdr.Obj.Width + BOOT_ACTION_BUTTON_SPACE * GuiContext->Scale;
-  mBootPickerRestart.Hdr.Obj.OffsetY = 0;
-
-  mBootPickerActionButtonsContainer.Obj.Width = mBootPickerShutDown.Hdr.Obj.Width + mBootPickerRestart.Hdr.Obj.Width + BOOT_ACTION_BUTTON_SPACE * GuiContext->Scale;
-  mBootPickerActionButtonsContainer.Obj.Height = MAX (mBootPickerShutDown.CurrentImage->Height, mBootPickerRestart.CurrentImage->Height);
+  mBootPickerActionButtonsContainer.Obj.Width =  mBootPickerRestart.Hdr.Obj.Width + mBootPickerShutDown.Hdr.Obj.Width + BOOT_ACTION_BUTTON_SPACE * GuiContext->Scale;
+  mBootPickerActionButtonsContainer.Obj.Height = MAX (mBootPickerRestart.CurrentImage->Height, mBootPickerShutDown.CurrentImage->Height);
   mBootPickerActionButtonsContainer.Obj.OffsetX = (mBootPickerView.Width - mBootPickerActionButtonsContainer.Obj.Width) / 2;
   mBootPickerActionButtonsContainer.Obj.OffsetY = mBootPickerView.Height - mBootPickerActionButtonsContainer.Obj.Height - BOOT_ACTION_BUTTON_SPACE * GuiContext->Scale;
 
@@ -1893,7 +1897,6 @@ BootPickerViewLateInitialize (
   INT64                  ScrollOffset;
   CONST LIST_ENTRY       *ListEntry;
   CONST GUI_VOLUME_ENTRY *BootEntry;
-  INT64                  FirstPosOffset;
 
   ASSERT (mBootPicker.SelectedEntry != NULL);
 
@@ -1904,30 +1907,37 @@ BootPickerViewLateInitialize (
   // impossible.
   //
   if (ScrollOffset == 0) {
-    ListEntry = mBootPicker.Hdr.Obj.Children.BackLink;
-    ASSERT (ListEntry == &mBootPickerSelector.Hdr.Link);
-
-    FirstPosOffset = 0;
-    //
-    // Last entry is always the selector.
-    //
-    ListEntry = ListEntry->BackLink;
     //
     // Find the first entry that is fully visible.
+    // Last entry is always the selector.
     //
-    while (!IsNull (&mBootPicker.Hdr.Obj.Children, ListEntry)) {
+    ASSERT (mBootPicker.Hdr.Obj.Children.BackLink == &mBootPickerSelector.Hdr.Link);
+    for (
+      ListEntry = GetFirstNode (&mBootPicker.Hdr.Obj.Children);
+      ListEntry != mBootPicker.Hdr.Obj.Children.BackLink;
+      ListEntry = GetNextNode (&mBootPicker.Hdr.Obj.Children, ListEntry)
+      ) {
+      //
+      // Move the first partially visible boot entry to the very left to prevent
+      // cut-off entries. This only applies when entries overflow.
+      //
       BootEntry = BASE_CR (ListEntry, GUI_VOLUME_ENTRY, Hdr.Link);
-      if (mBootPicker.Hdr.Obj.OffsetX + BootEntry->Hdr.Obj.OffsetX < 0) {
+      if (mBootPicker.Hdr.Obj.OffsetX + BootEntry->Hdr.Obj.OffsetX >= 0) {
         //
-        // Move the first fully visible boot entry to the very left to prevent
-        // cut-off entries. This only applies when entries overflow.
+        // Move the cut-off entry on-screen.
         //
-        ScrollOffset = -(INT64) FirstPosOffset;
+        ScrollOffset = -ScrollOffset;
         break;
       }
 
-      FirstPosOffset = mBootPicker.Hdr.Obj.OffsetX + BootEntry->Hdr.Obj.OffsetX;
-      ListEntry = ListEntry->BackLink;
+      ScrollOffset = mBootPicker.Hdr.Obj.OffsetX + BootEntry->Hdr.Obj.OffsetX;
+    }
+
+    if (mBootPicker.Hdr.Obj.Children.BackLink != mBootPicker.Hdr.Obj.Children.ForwardLink) {
+      //
+      // mBootPicker must not be entirely off-screen.
+      //
+      ASSERT (ListEntry != mBootPicker.Hdr.Obj.Children.BackLink);
     }
   }